### Description

Follow-up to #11830 . These need to be filtered, to avoid pulling repeatedly.