@bchocho 2024-03-05
I don't see a huge difference when running with the block limit, but I'll land as the change logically is what we want.

```
With fix:

Workload                                 | submitted/s  | committed/s  | expired/s    | rejected/s   | chain txn/s  | latency      | p50 lat      | p90 lat      | p99 lat      | batch->pos   | pos->prop    | prop->order  | order->commit | actual dur  
0: AccountGeneration[100.0k]             | 5767         | 5767         | 0            | 0            | 5786         | 17338        | 17200        | 17900        | 18800        | 0.206        | 0.149        | 0.511        | 0.464        | 198         
1: VectorPictureRead40[50.0k]            | 4716         | 4716         | 0            | 0            | 4730         | 10560        | 9200         | 15500        | 19800        | 0.204        | 0.138        | 0.512        | 0.371        | 198         
2: VectorPicture30k[10.0k]               | 216          | 119          | 96           | 0            | 121          | 27956        | 27400        | 54600        | 58700        | 0.205        | 0.193        | 0.763        | 1.064        | 198         
continuous traffic                       | submitted/s  | committed/s  | expired/s    | rejected/s   | latency      | p50 lat      | p90 lat      | p99 lat     
continuous with traffic 0                | 10           | 10           | 0            | 0            | 15553        | 17200        | 17800        | 18400       
continuous with traffic 1                | 10           | 10           | 0            | 0            | 9402         | 9300         | 15000        | 15600       
continuous with traffic 2                | 10           | 10           | 0            | 0            | 2574         | 2700         | 3300         | 3600        
Test Ok


Without fix:

Workload                                 | submitted/s  | committed/s  | expired/s    | rejected/s   | chain txn/s  | latency      | p50 lat      | p90 lat      | p99 lat      | batch->pos   | pos->prop    | prop->order  | order->commit | actual dur  
0: AccountGeneration[100.0k]             | 5859         | 5859         | 0            | 0            | 5877         | 17066        | 17000        | 17400        | 17900        | 0.206        | 0.118        | 0.492        | 0.417        | 198         
1: VectorPictureRead40[50.0k]            | 4852         | 4852         | 0            | 0            | 4855         | 10276        | 9100         | 13900        | 16700        | 0.205        | 0.118        | 0.499        | 0.354        | 198         
2: VectorPicture30k[10.0k]               | 211          | 110          | 100          | 0            | 115          | 28620        | 28600        | 55000        | 59400        | 0.218        | 0.282        | 0.803        | 1.050        | 198         
continuous traffic                       | submitted/s  | committed/s  | expired/s    | rejected/s   | latency      | p50 lat      | p90 lat      | p99 lat     
continuous with traffic 0                | 9            | 9            | 0            | 0            | 15508        | 16900        | 17500        | 18300       
continuous with traffic 1                | 10           | 10           | 0            | 0            | 9118         | 9000         | 12600        | 14400       
continuous with traffic 2                | 10           | 10           | 0            | 0            | 2792         | 3000         | 3600         | 4200        
Test Ok
```